refactor: move problem into util/.
* i didn't have a huge problem with where it was but the present effort
  comes so close to completely cleaning up the root dir that i'd rather
  just do it.
* a minor argument could be made for outbound/ but it neither templates
  anything nor does any io itself so it doesn't seem to actually work.
