+++ lib/bin/backup.js
@@ -7,7 +7,7 @@ const config = require('config');
 const tmp = require('tmp');
 const archiver = require('archiver');
 const { google } = require('googleapis');
-const Problem = require('../problem');
+const Problem = require('../util/problem');
 const Option = require('../util/option');
 const { connect } = require('../model/database');
 const { getOrElse } = require('../util/http');
