+++ lib/model/instance/actor.js
@@ -1,6 +1,6 @@
 const Instance = require('./instance');
 const ActeeTrait = require('../trait/actee');
-const Problem = require('../../problem');
+const Problem = require('../../util/problem');
 const { reject } = require('../../util/promise');
 const { withCreateTime, withUpdateTime } = require('../../util/instance');
 const { ensureArray } = require('../../util/util');
