+++ lib/model/instance/auth.js
@@ -1,6 +1,6 @@
 const Instance = require('./instance');
 const { reject } = require('../../util/promise');
-const Problem = require('../../problem');
+const Problem = require('../../util/problem');
 
 module.exports = Instance(({ Grant }) => class {
   // See: Grant::can for behaviour. TODO: also see it for todo notes.
