+++ lib/model/instance/form.js
@@ -4,7 +4,7 @@ const { createHash } = require('crypto');
 const { toTraversable, findAndTraverse, traverseFirstChild } = require('../../util/xml');
 const Instance = require('./instance');
 const ActeeTrait = require('../trait/actee');
-const Problem = require('../../problem');
+const Problem = require('../../util/problem');
 const { isBlank } = require('../../util/util');
 const { withCreateTime } = require('../../util/instance');
 const Option = require('../../util/option');
