+++ lib/model/instance/submission.js
@@ -3,7 +3,7 @@ const { validate } = require('fast-xml-parser');
 const uuid = require('uuid/v4');
 const { toTraversable, findAndTraverse, traverseFirstChild } = require('../../util/xml');
 const Instance = require('./instance');
-const Problem = require('../../problem');
+const Problem = require('../../util/problem');
 const { isBlank } = require('../../util/util');
 const { withCreateTime } = require('../../util/instance');
 const Option = require('../../util/option');
