+++ lib/model/query/actors.js
@@ -1,4 +1,4 @@
-const Problem = require('../../problem');
+const Problem = require('../../util/problem');
 const { getOrReject } = require('../../util/http'); // TODO: clearly, this is useful beyond http.
 
 module.exports = {
