+++ lib/model/query/submissions.js
@@ -1,4 +1,4 @@
-const Problem = require('../../problem');
+const Problem = require('../../util/problem');
 const { fieldsForJoin, joinRowToInstance, rowsToInstances, maybeFirst, translateProblem } = require('../../util/db');
 const { resolve } = require('../../util/promise');
