+++ lib/resources/config.js
@@ -2,7 +2,7 @@ const { google } = require('googleapis');
 const config = require('config');
 const { isBlank, printPairs } = require('../util/util');
 const { endpoint, getOrNotFound, getOrElse, success } = require('../util/http');
-const Problem = require('../problem');
+const Problem = require('../util/problem');
 const Option = require('../util/option');
 const { ExplicitPromise, reject } = require('../util/promise');
 const { generateKeypair } = require('../util/crypto');
