+++ lib/resources/sessions.js
@@ -1,4 +1,4 @@
-const Problem = require('../problem');
+const Problem = require('../util/problem');
 const { isBlank } = require('../util/util');
 const { verifyPassword } = require('../util/crypto');
 const { endpoint, getOrReject, success } = require('../util/http');
