+++ lib/resources/submissions.js
@@ -5,7 +5,7 @@ const { openRosaEndpoint, endpoint, getOrNotFound, getOrReject, rejectIf } = req
 const { openRosaMessage } = require('../outbound/openrosa');
 const { resolve, reject } = require('../util/promise');
 const Option = require('../util/option');
-const Problem = require('../problem');
+const Problem = require('../util/problem');
 const { streamJoinedCsvs } = require('../data/csv');
 const { streamAttachments } = require('../data/attachments');
 const { zipStreamFromParts } = require('../util/zip');
