+++ lib/resources/users.js
@@ -1,6 +1,6 @@
 const { endpoint, getOrNotFound, success, isTrue } = require('../util/http');
 // const { verifyPassword } = require('../util/crypto');
-const Problem = require('../problem');
+const Problem = require('../util/problem');
 const { resolve, reject } = require('../util/promise');
 
 module.exports = (service, { Actee, User, mail }) => {
