+++ lib/util/db.js
@@ -1,6 +1,6 @@
 const { inspect } = require('util');
 const { reject } = require('../util/promise');
-const Problem = require('../problem');
+const Problem = require('../util/problem');
 const Option = require('../util/option');
