+++ lib/util/http.js
@@ -4,7 +4,7 @@ const { inspect } = require('util');
 const { DateTime } = require('luxon');
 const { isBlank } = require('./util');
 const { openRosaMessage } = require('../outbound/openrosa');
-const Problem = require('../problem');
+const Problem = require('../util/problem');
 const Option = require('../util/option');
 const { reject } = require('../util/promise');
