+++ lib/util/problem.js
@@ -1,5 +1,5 @@
 const { floor } = Math;
-const { printPairs } = require('./util/util');
+const { printPairs } = require('./util');
 
 // Problem is our generic catch-all for all things that may go wrong in the
 // system, and any layer. Our infrastructure is set up to automatically interpret
